Refactored Discord event interfaces to distinguish between direct and scheduled synchronous handling
